FlaskWebProject2/requirements.txt
appdirs==1.4.4
Babel==2.9.0
click==7.1.2
distlib==0.3.1
dnspython==2.0.0
filelock==3.0.12
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Table==0.5.0
gunicorn==20.0.4
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymongo==3.11.2
pytz==2020.4
six==1.15.0
typing-extensions==3.7.4.3
virtualenv==20.2.2
Werkzeug==1.0.1
zipp==3.4.0
